+++ source/features/fault_models/temperature/adiabatic.cc
@@ -127,7 +127,8 @@ namespace WorldBuilder
                                    double temperature_,
                                    const double  /*feature_min_depth*/,
                                    const double  /*feature_max_depth*/,
-                                   const WorldBuilder::Utilities::PointDistanceFromCurvedPlanes & /*distance_from_planes*/) const
+                                   const WorldBuilder::Utilities::PointDistanceFromCurvedPlanes & /*distance_from_planes*/,
+                                   const Utilities::AdditionalParameters & /*additional_paramters*/) const
         {
 
           if (depth <= max_depth && depth >= min_depth)