+++ source/features/subducting_plate.cc
@@ -331,15 +331,6 @@ namespace WorldBuilder
           maximum_total_slab_length = std::max(maximum_total_slab_length, local_total_slab_length);
         }
 
-      std::vector<double> x_list(original_number_of_coordinates,0.0);
-      std::vector<double> y_list(original_number_of_coordinates,0.0);
-
-      for (size_t j=0; j<original_number_of_coordinates; ++j)
-        {
-          x_list[j] = coordinates[j][0];
-          y_list[j] = coordinates[j][1];
-        }
-
       // Here, we compute the spherical bounding box using the two extreme points of the box containing all the surface
       // coordinates and an additional buffer zone that accounts for the fault thickness and length. The first and second
       // points correspond to the lower left and the upper right corners of the bounding box, respectively (see the
@@ -347,16 +338,33 @@ namespace WorldBuilder
       // For the spherical system, the buffer zone along the longitudal direction is calculated using the
       // correponding latitude points.
 
-      min_along_x = *std::min_element(x_list.begin(), x_list.end());
-      max_along_x = *std::max_element(x_list.begin(), x_list.end());
-      min_along_y = *std::min_element(y_list.begin(), y_list.end());
-      max_along_y = *std::max_element(y_list.begin(), y_list.end());
+      // Find minimal and maximal coordinates. Do this by finding the
+      // leftmost/rightmost point with regard to either the [0] or [1]
+      // coordinate, and then takes its [0] or [1] element.
+      auto compare_x_coordinate = [](auto p1, auto p2)
+      {
+        return p1[0]<p2[0];
+      };
 
-      min_lat_cos_inv = 1 / std::cos(min_along_y);
-      max_lat_cos_inv = 1 / std::cos(max_along_y);
+      min_along_x = (*std::min_element(coordinates.begin(), coordinates.end(), compare_x_coordinate)) [0];
+      max_along_x = (*std::max_element(coordinates.begin(), coordinates.end(), compare_x_coordinate)) [0];
 
-      buffer_around_slab_cartesian =  (maximum_slab_thickness + maximum_total_slab_length);
+
+      auto compare_y_coordinate = [](auto p1, auto p2)
+      {
+        return p1[1]<p2[1];
+      };
+
+      min_along_y = (*std::min_element(coordinates.begin(), coordinates.end(), compare_y_coordinate)) [1];
+      max_along_y = (*std::max_element(coordinates.begin(), coordinates.end(), compare_y_coordinate)) [1];
+
+      min_lat_cos_inv = 1. / std::cos(min_along_y);
+      max_lat_cos_inv = 1. / std::cos(max_along_y);
+
+      buffer_around_slab_cartesian = (maximum_slab_thickness + maximum_total_slab_length);
     }
+
+
 
     BoundingBox<2>
     SubductingPlate::get_bounding_box (const NaturalCoordinate &position_in_natural_coordinates,
@@ -460,7 +468,6 @@ namespace WorldBuilder
                                             * (slab_segment_thickness[next_section][current_segment][1]
                                                - slab_segment_thickness[current_section][current_segment][1]);
               const double thickness_local = thickness_up + segment_fraction * (thickness_down - thickness_up);
-              distance_from_planes.local_thickness = thickness_local;
 
               // secondly for top truncation
               const double top_truncation_up = slab_segment_top_truncation[current_section][current_segment][0]
@@ -491,6 +498,7 @@ namespace WorldBuilder
                   distance_along_plane <= max_slab_length)
                 {
                   // Inside the slab!
+                  const Features::Utilities::AdditionalParameters additional_parameters = {max_slab_length,thickness_local};
                   double temperature_current_section = temperature;
                   double temperature_next_section = temperature;
 
@@ -502,7 +510,8 @@ namespace WorldBuilder
                                                                                        temperature_current_section,
                                                                                        starting_depth,
                                                                                        maximum_depth,
-                                                                                       distance_from_planes);
+                                                                                       distance_from_planes,
+                                                                                       additional_parameters);
 
                       WBAssert(!std::isnan(temperature_current_section), "Temparture is not a number: " << temperature_current_section
                                << ", based on a temperature model with the name " << temperature_model->get_name());
@@ -519,7 +528,8 @@ namespace WorldBuilder
                                                                                     temperature_next_section,
                                                                                     starting_depth,
                                                                                     maximum_depth,
-                                                                                    distance_from_planes);
+                                                                                    distance_from_planes,
+                                                                                    additional_parameters);
 
                       WBAssert(!std::isnan(temperature_next_section), "Temparture is not a number: " << temperature_next_section
                                << ", based on a temperature model with the name " << temperature_model->get_name());
@@ -590,7 +600,6 @@ namespace WorldBuilder
                                             * (slab_segment_thickness[next_section][current_segment][1]
                                                - slab_segment_thickness[current_section][current_segment][1]);
               const double thickness_local = thickness_up + segment_fraction * (thickness_down - thickness_up);
-              distance_from_planes.local_thickness = thickness_local;
 
               // secondly for top truncation
               const double top_truncation_up = slab_segment_top_truncation[current_section][current_segment][0]
@@ -621,6 +630,7 @@ namespace WorldBuilder
                   distance_along_plane <= max_slab_length)
                 {
                   // Inside the slab!
+                  const Features::Utilities::AdditionalParameters additional_parameters = {max_slab_length,thickness_local};
                   double composition_current_section = composition;
                   double composition_next_section = composition;
 
@@ -632,7 +642,8 @@ namespace WorldBuilder
                                                                                        composition_current_section,
                                                                                        starting_depth,
                                                                                        maximum_depth,
-                                                                                       distance_from_planes);
+                                                                                       distance_from_planes,
+                                                                                       additional_parameters);
 
                       WBAssert(!std::isnan(composition_current_section), "Composition_current_section is not a number: " << composition_current_section
                                << ", based on a temperature model with the name " << composition_model->get_name());
@@ -649,7 +660,8 @@ namespace WorldBuilder
                                                                                     composition_next_section,
                                                                                     starting_depth,
                                                                                     maximum_depth,
-                                                                                    distance_from_planes);
+                                                                                    distance_from_planes,
+                                                                                    additional_parameters);
 
                       WBAssert(!std::isnan(composition_next_section), "Composition_next_section is not a number: " << composition_next_section
                                << ", based on a temperature model with the name " << composition_model->get_name());
@@ -723,7 +735,6 @@ namespace WorldBuilder
                                             * (slab_segment_thickness[next_section][current_segment][1]
                                                - slab_segment_thickness[current_section][current_segment][1]);
               const double thickness_local = thickness_up + segment_fraction * (thickness_down - thickness_up);
-              distance_from_planes.local_thickness = thickness_local;
 
               // secondly for top truncation
               const double top_truncation_up = slab_segment_top_truncation[current_section][current_segment][0]
@@ -754,6 +765,7 @@ namespace WorldBuilder
                   distance_along_plane <= max_slab_length)
                 {
                   // Inside the slab!
+                  const Features::Utilities::AdditionalParameters additional_parameters = {max_slab_length,thickness_local};
                   WorldBuilder::grains  grains_current_section = grains;
                   WorldBuilder::grains  grains_next_section = grains;
 
@@ -765,7 +777,8 @@ namespace WorldBuilder
                                                                         grains_current_section,
                                                                         starting_depth,
                                                                         maximum_depth,
-                                                                        distance_from_planes);
+                                                                        distance_from_planes,
+                                                                        additional_parameters);
 
                       /*WBAssert(!std::isnan(composition_current_section), "Composition_current_section is not a number: " << composition_current_section
                                << ", based on a temperature model with the name " << composition_model->get_name());
@@ -782,7 +795,8 @@ namespace WorldBuilder
                                                                      grains_next_section,
                                                                      starting_depth,
                                                                      maximum_depth,
-                                                                     distance_from_planes);
+                                                                     distance_from_planes,
+                                                                     additional_parameters);
 
                       /*WBAssert(!std::isnan(composition_next_section), "Composition_next_section is not a number: " << composition_next_section
                                << ", based on a temperature model with the name " << composition_model->get_name());